+++ src/core/index.ts
@@ -8,4 +8,5 @@ export * from "./parse";
 export * from "./colorConfig";
 export * from "./titleSettings";
 export * from "./visibility";
+export * from "./namingRules";
 export * from "./config";